requirements.ignore.txt
absl-py==0.6.1
argh==0.26.2
astor==0.7.1
astroid==2.1.0
atari-py==0.1.7
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
backcall==0.1.0
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
cloudpickle==0.6.1
cxxfilt==0.2.0
cycler==0.10.0
cymem==2.0.2
Cython==0.29
decorator==4.3.0
fancycompleter==0.8
future==0.16.0
gast==0.2.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-logging==1.10.0
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
grpcio==1.17.1
gym==0.10.8
httplib2==0.12.0
idna==2.7
ipdb==0.11
ipython==7.0.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.1
jsmin==2.2.2
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
matplotlib==3.0.0
mccabe==0.6.1
mlperf-compliance==0.0.6
more-itertools==4.3.0
mpi4py==3.0.0
numpy==1.14.5
oauth2client==4.1.0
opencv-python==3.4.4.19
pandas==0.23.4
parso==0.3.1
pathlib2==2.3.2
pdbpp==0.9.3
petname==2.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pluggy==0.8.0
progressbar==2.5
progressbar2==3.38.0
prompt-toolkit==2.0.6
protobuf==3.6.1
psutil==5.4.7
psycopg2==2.7.7
ptyprocess==0.6.0
py==1.7.0
py-cpuinfo==4.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycodestyle==2.4.0
pyglet==1.3.2
Pygments==2.2.0
pygtp==0.4
pylint==2.2.2
PyOpenGL==3.1.0
pyparsing==2.2.2
pytest==4.0.0
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.6
requests==2.20.0
rsa==4.0
scipy==1.1.0
scons==3.0.1
seaborn==0.9.0
sgf==0.5
simplegeneric==0.8.1
six==1.12.0
tensorboard==1.10.0
-e git+git@github.com:jagleeso/tensorflow.benchmark_c_api.git@23dd8f54392eff95929c02541eba1bee62930dc0#egg=tensorflow&subdirectory=_python_build
termcolor==1.1.0
tqdm==4.29.1
traitlets==4.3.2
typed-ast==1.2.0
urllib3==1.24
wcwidth==0.1.7
Werkzeug==0.14.1
wmctrl==0.3
wrapt==1.11.1
